+++ lib/capybara.rb
@@ -406,6 +406,7 @@ module Capybara
   require 'capybara/queries/match_query'
   require 'capybara/queries/ancestor_query'
   require 'capybara/queries/sibling_query'
+  require 'capybara/queries/style_query'
 
   require 'capybara/node/finders'
   require 'capybara/node/matchers'