+++ lib/capybara/rack_test/node.rb
@@ -24,7 +24,7 @@ class Capybara::RackTest::Node < Capybara::Driver::Node
     string_node[name]
   end
 
-  def style(styles)
+  def style(_styles)
     raise NotImplementedError, "The rack_test driver does not process CSS"
   end